requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pydantic
python-dotenv